Remove zephemeral capabilities from codebase
The ro-root.sh file that allowed for the zephemeral capabilities inside
of the zram-config project is licensed under the terms of the GPLv3
license. As the GPLv3 license cannot be used under the MIT license,
which is the license that the zram-config project is distributed under,
the use of the ro-root.sh file in the zram-config project violated the
terms of the GPLv3 license and as such could not legally be distributed
with the zram-config project. To prevent potential legal action against
the zram-config project, the ro-root.sh file and the capabilities it
provided in the zram-config project are removed by this commit.

Signed-off-by: Ethan Dye <mrtops03@gmail.com>