Allow sync to persistent storage
Using the same merge functionality in overlayfs-tools already used to
write changes to persistent storage it is possible to write changes to
disk without completely reseting the zram device. This allows for a
quick sync to persistent storage to be run whenever desired.

Without the ability to sync, any changes made to the underlying
filesystem will be lost if there was a sudden power loss or system
shutdown. This can be mitigated by periodically running the sync command
which will write the changes to persistent storage ensuring that the
data will not be lost if anything goes wrong.

A default service that runs the sync every night is included in this
change, however, it is not installed by default.

Fixes #72

Signed-off-by: Ethan Dye <mrtops03@gmail.com>