Fix tests broken by upstream changes
Raspberry Pi OS recently switched to a new image compression method of
XZ. This resolves the incompatibility with the new form of compression.

Additionally, Raspberry Pi OS no longer ships with a default username
and password and as such we now need to create a username and password
manually in order to run our tests.

Signed-off-by: Ethan Dye <mrtops03@gmail.com>