Ensure that files sync to disk on SystemD (#123)
This ensures that zram-config has the correct dependency ordering to
startup and shutdown correctly. In this case, making it start up as
early as possible so we are ready before anyone starts writing to our
directories. This includes starting up before rsyslog and
systemd-journald.

An added conflict with shutdown targets is required to make sure that
the service stops before shutdown when DefaultDependencies is disabled.

IgnoreOnIsolate simply prevents taking the service down when isolating
other units.

Signed-off-by: Ethan Dye <mrtops03@gmail.com>